+++ client-api/src/main/java/com/hyf/hotrefresh/client/api/core/DefaultRequestBuilder.java
@@ -1,5 +1,8 @@
 package com.hyf.hotrefresh.client.api.core;
 
+import com.hyf.hotrefresh.common.Constants;
+import com.hyf.hotrefresh.common.Log;
+import com.hyf.hotrefresh.common.Services;
 import com.hyf.hotrefresh.remoting.constants.RemotingConstants;
 import com.hyf.hotrefresh.remoting.message.Message;
 import com.hyf.hotrefresh.remoting.message.MessageCodec;
@@ -10,6 +13,12 @@ import org.apache.http.entity.ContentType;
 import org.apache.http.entity.InputStreamEntity;
 
 import java.io.ByteArrayInputStream;
+import java.io.UnsupportedEncodingException;
+import java.net.URLEncoder;
+import java.util.HashMap;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Map;
 
 /**
  * @author baB_hyf
@@ -19,6 +28,13 @@ public class DefaultRequestBuilder implements RequestBuilder {
 
     public static final RequestConfig DEFAULT_CONFIG = createDefaultConfig();
 
+    private final List<RequestInterceptor> interceptors;
+
+    public DefaultRequestBuilder() {
+        interceptors = Services.gets(RequestInterceptor.class);
+    }
+
+    // TODO config
     private static RequestConfig createDefaultConfig() {
         return RequestConfig.custom().setConnectionRequestTimeout(1000)
                 .setConnectTimeout(10000).setSocketTimeout(60000).build();
@@ -27,12 +43,48 @@ public class DefaultRequestBuilder implements RequestBuilder {
     @Override
     public HttpUriRequest build(String url, Message message) {
 
-        HttpPost post = new HttpPost(url);
+        HttpRequest request = new HttpRequest();
+        request.setUrl(url);
+
+        interceptors.forEach(interceptor -> interceptor.intercept(request));
+
+        HttpPost post = new HttpPost(getUrl(request));
         post.setConfig(DEFAULT_CONFIG);
+        request.getHeaders().forEach(post::addHeader);
 
         ByteArrayInputStream bais = new ByteArrayInputStream(MessageCodec.encode(message));
         post.setEntity(new InputStreamEntity(bais, bais.available(), ContentType.create(RemotingConstants.DEFAULT_CONTENT_TYPE)));
 
         return post;
+    }
+
+    private String getUrl(HttpRequest request) {
+        StringBuilder sb = new StringBuilder(request.getUrl());
+
+        Map<String, String> params = request.getParams();
+
+        if (!params.isEmpty()) {
+            boolean urlHasParams = sb.indexOf("?") != -1;
+            if (!urlHasParams) {
+                sb.append('?');
+            }
+            else {
+                sb.append('&');
+            }
+            List<String> kvPairs = new LinkedList<>();
+            params.forEach((k, v) -> kvPairs.add(urlEncode(k) + "=" + urlEncode(v)));
+            sb.append(String.join("&", kvPairs));
+        }
+
+        return sb.toString();
+    }
+
+    private String urlEncode(String s) {
+        try {
+            return URLEncoder.encode(s, Constants.MESSAGE_ENCODING.toString());
+        } catch (UnsupportedEncodingException e) {
+            Log.error("Url encode failed", e);
+            return s;
+        }
     }
 }